[feature/spacex] [feature/spacex] Update SpaceX data layer
