+++ feature/spacex/src/main/java/com/alxnophis/jetpack/spacex/data/datasource/LaunchesDataSourceImpl.kt
@@ -3,7 +3,7 @@ package com.alxnophis.jetpack.spacex.data.datasource
 import arrow.core.Either
 import com.alxnophis.jetpack.api.spacex.SpacexApi
 import com.alxnophis.jetpack.kotlin.utils.DispatcherProvider
-import com.alxnophis.jetpack.spacex.data.datasource.mapper.map
+import com.alxnophis.jetpack.spacex.data.model.mapper.map
 import com.alxnophis.jetpack.spacex.data.model.PastLaunchesDataModel
 import com.alxnophis.jetpack.spacex.data.model.LaunchesError
 import com.apollographql.apollo3.exception.ApolloHttpException
